user done, fixes

+++ Controllers/EquipmentController.cs
@@ -19,34 +19,87 @@ namespace gutv_booker.Controllers
         [HttpPost("createtype")]
         public async Task<ActionResult<EquipmentType>> CreateEquipmentType([FromBody] EquipmentType equipmentType)
         {
-            var eqType = await _equipmentService.CreateEquipmentType(equipmentType.Name, equipmentType.Description, equipmentType.Category, equipmentType.AttributesJson);
-            return Ok(eqType);
+            if (string.IsNullOrEmpty(equipmentType.Name))
+            {
+                return BadRequest("Название оборудования не может быть пустым");
+            }
+
+            if (!Enum.IsDefined(typeof(EquipmentType.EquipmentCategory), equipmentType.Category))
+                return BadRequest("Некорректная категория оборудования");
+
+            try
+            {
+                var eqType = await _equipmentService.CreateEquipmentType(equipmentType.Name, equipmentType.Description,
+                    equipmentType.Category, equipmentType.AttributesJson);
+                return Ok(eqType);
+            }
+
+            catch (InvalidOperationException ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
-        // DELETE api/equipment/delete/{id}
-        [HttpDelete("delete/{id}")]
-        public async Task<ActionResult> DeleteEquipmentType(int id)
+        // GET api/equipment/get_all_types
+        [HttpGet("get_all_types")]
+        public async Task<ActionResult<List<EquipmentType>>> GetAllTypes()
         {
-            var succes = await _equipmentService.DeleteEquipmentType(id);
-            if (!succes) return NotFound();
-            return Ok();
+            var eqTypes = await _equipmentService.GetAllEquipmentTypes();
+            if (!eqTypes.Any())
+                return NotFound("Оборудование не найдено");
+
+            return Ok(eqTypes);
         }
 
         // PUT api/equipment/update/{id}
         [HttpPut("update/{id}")]
         public async Task<ActionResult> UpdateEquipmentType(int id, [FromBody] EquipmentType equipmentType)
         {
-            var success = await _equipmentService.UpdateEquipmentType(
-                id,
-                name: equipmentType.Name,
-                description: equipmentType.Description,
-                category: equipmentType.Category,
-                attributesJson: equipmentType.AttributesJson
-            );
+            if (id <= 0)
+                return BadRequest("ID должен быть больше нуля");
 
-            if (!success) return NotFound();
-            return Ok();
+            if (string.IsNullOrWhiteSpace(equipmentType.Name))
+                return BadRequest("Название оборудования не может быть пустым");
+
+            if (!Enum.IsDefined(typeof(EquipmentType.EquipmentCategory), equipmentType.Category))
+                return BadRequest("Категория оборудования некорректна");
+
+            var success = await _equipmentService.UpdateEquipmentType(id, name: equipmentType.Name, description: equipmentType.Description, category: equipmentType.Category, attributesJson: equipmentType.AttributesJson);
+
+            if (!success) return NotFound($"Оборудование с ID {id} не найдено");
+            return Ok("Обновление прошло успешно");
         }
 
+        // DELETE api/equipment/delete/{id}
+        [HttpDelete("delete/{id}")]
+        public async Task<ActionResult> DeleteEquipmentType(int id)
+        {
+            if (id <= 0)
+                return BadRequest("ID должен быть больше нуля");
+
+            var success = await _equipmentService.DeleteEquipmentType(id);
+            if (!success) return NotFound($"Оборудование с ID {id} не найдено");
+
+            return Ok("Удаление прошло успешно");
+        }
+
+
+        // POST api/equipment/createitem
+        [HttpPost("createitem")]
+        public async Task<ActionResult<EquipmentItem>> CreateEquipmentItem([FromBody] EquipmentItem equipmentItem)
+        {
+            if (equipmentItem.EquipmentTypeId <= 0)
+                return BadRequest("Некорректный EquipmentTypeId");
+
+            try
+            {
+                var createdItem = await _equipmentService.CreateEquipmentItem(equipmentItem.EquipmentTypeId, equipmentItem.Available);
+                return Ok(createdItem);
+            }
+            catch (Exception ex)
+            {
+                return NotFound(ex.Message);
+            }
+        }
     }
 }
